+++ src/main/java/com/devbliss/gpullr/service/github/GithubApi.java
@@ -239,7 +239,7 @@ public class GithubApi {
     pullRequest.linesRemoved = pullRequestJson.getInt("deletions");
     pullRequest.filesChanged = pullRequestJson.getInt("changed_files");
     pullRequest.number = pullRequestJson.getInt("number");
-    pullRequest.numberOfReviewComments = pullRequestJson.getInt("review_comments");
+    pullRequest.numberOfComments = pullRequestJson.getInt("review_comments") + pullRequestJson.getInt("comments");
     JsonValue assigneeValue = pullRequestJson.get(FIELD_KEY_ASSIGNEE);
 
     // unfortunately, assignee is only set when PR is CLOSED - so it's useless for us!