+++ src/test/java/com/devbliss/gpullr/controller/dto/PullRequestConverterUnitTest.java
@@ -72,7 +72,7 @@ public class PullRequestConverterUnitTest {
 
   private static final String ASSSIGNEE_PROFILE_URL = "http://assign.yourself.example.com";
 
-  private static final int NUMBER_OF_REVIEW_COMMENTS = 14;
+  private static final int NUMBER_OF_COMMENTS = 14;
 
   private static final BuildStatus BUILD_STATUS = new BuildStatus(
       BuildStatus.State.PENDING,
@@ -105,7 +105,7 @@ public class PullRequestConverterUnitTest {
     entity.url = URL;
     entity.assignedAt = ASSIGNED_AT;
     entity.buildStatus = BUILD_STATUS;
-    entity.numberOfReviewComments = NUMBER_OF_REVIEW_COMMENTS;
+    entity.numberOfComments = NUMBER_OF_COMMENTS;
   }
 
   @Test
@@ -128,7 +128,7 @@ public class PullRequestConverterUnitTest {
     assertEquals(TITLE, dto.title);
     assertEquals(URL, dto.url);
     assertEquals(ASSIGNED_AT_STRING, dto.assignedAt);
-    assertEquals(NUMBER_OF_REVIEW_COMMENTS, dto.numberOfComments.intValue());
+    assertEquals(NUMBER_OF_COMMENTS, dto.numberOfComments.intValue());
     assertEquals(BUILD_STATUS.state.name(), dto.buildStatus);
   }
 }