+++ changelog.md
@@ -2,6 +2,7 @@
 ### Version 0.1.0-SNAPSHOT
 
 
+ - [bug] number of comments in a pullrequest is now the sum of both normal and review comments
  - [bug] correct field for fetching number of review comments
  - [feature] number of comments and build status in pull request data
  - [feature] added user profile url to user
@@ -46,5 +47,5 @@
  - [feature] project setup
  - [initial] initial commit
 
--- Last change by: hschuetz <henning.schuetz@devbliss.com> Wed Apr 08 09:14:25 CEST 2015
+-- Last change by: hschuetz <henning.schuetz@devbliss.com> Wed Apr 08 11:01:41 CEST 2015